# go-core/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateStrategyRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      user_id:
        type: integer
    required:
    - description
    - name
    - user_id
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.GetStrategiesResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      strategies:
        items:
          $ref: '#/definitions/dto.StrategyResponse'
        type: array
    type: object
...
  dto.PaginationResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  dto.StrategyResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  dto.UpdateStrategyRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 1
        type: string
      id:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      user_id:
        type: integer
    required:
    - description
    - id
    - name
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive trading journal API for tracking trades, setups, and
    market analysis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 100xTrader API
  version: "1.0"
paths:
...
  /api/v1/strategies:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of strategies
      parameters:
      - description: 'Number of strategies to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of strategies to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Strategies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetStrategiesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List strategies
      tags:
      - strategies
    post:
      consumes:
      - application/json
      description: Create a new trading strategy
      parameters:
      - description: Strategy data
        in: body
        name: strategy
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStrategyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Strategy created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StrategyResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new strategy
      tags:
      - strategies
  /api/v1/strategies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific strategy by ID
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Strategy deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Invalid strategy ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Strategy not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a strategy
      tags:
      - strategies
    get:
      consumes:
      - application/json
      description: Retrieve a specific strategy with all its details
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Strategy retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StrategyResponse'
              type: object
        "400":
          description: Invalid strategy ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Strategy not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a strategy by ID
      tags:
      - strategies
    put:
      consumes:
      - application/json
      description: Update an existing strategy with new data
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated strategy data
        in: body
        name: strategy
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStrategyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Strategy updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StrategyResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Strategy not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a strategy
      tags:
      - strategies
  /api/v1/strategies/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of strategies for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of strategies to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of strategies to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User strategies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetStrategiesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get strategies by user
      tags:
      - strategies
...
schemes:
- http
swagger: "2.0"